// frontend/apps/stitch-admin/src/shared/api/generated/model/routeResponseDTO.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Stitch API Gateway
 * OpenAPI spec version: 1.0.0
 */
import type { StopResponseDTO } from './stopResponseDTO';

export interface RouteResponseDTO {
  id: string;
  execution_date: string;
  status: string;
  estimated_volume?: number | null;
  estimated_weight?: number | null;
  planned_distance?: number | null;
  planned_duration?: number | null;
  vehicle_id?: string | null;
  driver_id?: string | null;
  stops: StopResponseDTO[];
}